PHOENIX-4466 Relocate Avatica and hadoop-common in thin-client jar (Toshihiro Suzuki)
When using the thin-client in Spark, we encounter problems in that Spark
is placing its own version of avatica on the classpath as well. We can
relocate most of Avatica (all but the protobuf generated messages as
their classnames are required to be 'org.apache.calcite.avatica.proto'
presently) and hadoop-common to avoid future problems.

(cherry picked from commit 4da5c80ad26f7671a7f80a6e34b35988082bbee2)

Change-Id: Iacb9cf97313e9afb70c682c3e5fe6dd604501c75